requirements.txt
anaconda-client==1.7.2
anaconda-navigator==2.0.3
aniso8601==8.0.0
appdirs==1.4.4
applaunchservices==0.2.1
argh==0.26.2
asgiref==3.5.2
atomicwrites==1.4.0
backports.weakref==1.0.post1
backports.zoneinfo==0.2.1
bkcharts==0.2
black==19.10b0
boto==2.49.0
Bottleneck==1.3.2
brotlipy==0.7.0
certifi==2020.12.5
clyent==1.2.2
conda==4.10.1
conda-build==3.21.4
conda-verify==3.4.2
contextlib2==0.6.0.post1
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
cytoolz==0.11.0
distlib==0.3.4
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.7.0
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
future==0.18.2
gmpy2==2.0.8
h5py==2.10.0
HeapDict==1.0.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
json5==0.9.5
llvmlite==0.36.0
locket==0.2.1
mccabe==0.6.1
mkl-fft==1.3.0
mkl-service==2.3.0
mpmath==1.2.1
multipledispatch==0.6.0
mypy-extensions==0.4.3
navigator-updater==0.2.1
olefile==0.46
pandas==1.2.4
parso==0.7.0
pathspec==0.7.0
patsy==0.5.1
pep8==1.7.1
pkginfo==1.7.0
platformdirs==2.5.2
ply==3.11
pycosat==0.6.3
pycurl==7.43.0.6
pymongo==3.11.4
pyodbc===4.0.0-unsupported
pytest==6.2.3
pytz==2020.1
PyYAML==5.4.1
pyzmq==20.0.0
QDarkStyle==2.8.1
QtPy==1.9.0
scikit-image==0.18.1
simplegeneric==0.8.1
SQLAlchemy==1.3.18
sqlparse==0.4.2
tables==3.6.1
terminado==0.9.4
tifffile==2020.10.1
unicodecsv==0.14.1
uritemplate==4.1.1
virtualenv==20.14.1
webencodings==0.5.1
widgetsnbextension==3.5.1
xlwings==0.23.0
xlwt==1.3.0
xmltodict==0.12.0
yellowbrick==1.2
zict==2.0.0
zope.event==4.5.0
